JavaScript: peaceful coexistence with Django 1.2's jQuery
Changes the default when running on Django 1.2 and higher to use the copy of
jQuery bundled with Django admin.

If Django < 1.2 or settings.USE_DJANGO_JQUERY = False jQuery will be loaded from
the Google CDN as before.

Note: as part of this change, jQuery usage is wrapped within a closure to avoid
touching the global namespace.